Split Client List form owner
