+++ tsconfig.checkjs.json
@@ -96,6 +96,7 @@
     "js/client-list-runtime.js",
     "js/client-list.js",
     "js/client-list-impl.js",
+    "js/client-list-form.js",
     "js/constants.js",
     "js/context-card-dashboard-ai-actions.js",
     "js/context-card-dashboard-ai-runtime.js",
