+++ tsconfig.json
@@ -93,6 +93,7 @@
     "js/client-list-runtime.js",
     "js/client-list.js",
     "js/client-list-impl.js",
+    "js/client-list-form.js",
     "js/changelog.js",
     "js/changelog-impl.js",
     "js/commit-hash.js",
